% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix.R
\name{admix}
\alias{admix}
\title{Constructor for admix object}
\usage{
admix()
}
\value{
an admix object which is a list with 6 elements:
 K: number of clusters estimated by ADMIXTURE


 nsamples: number of samples used

 nmarkers: number of markers used

 Q_df: a data.frame of cluster membership probabilities

 P_df: a data.frame of estimated marker frequencies in each inferred population

 log_info: a data.frame containing the K, CVerror and logLik of the last model.
}
\description{
Constructor for admix object
}

