% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix.R
\name{admixList}
\alias{admixList}
\title{Contructor for admixList}
\usage{
admixList(...)
}
\arguments{
\item{...}{a list of \code{\link{admix}} objects}
}
\value{
an admixList object
}
\description{
Collect many  \code{\link{admix}} objects
}

