% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmie-methods.R
\name{autoplot.starmie}
\alias{autoplot.starmie}
\title{ggplot method for starmie object}
\usage{
\method{autoplot}{starmie}(starmie_obj, method, type = "all", ...)
}
\arguments{
\item{starmie_obj}{an object of class \code{\link{starmie}}}

\item{method}{a character corresponding to either 'admixture' or 'structure'}

\item{type}{a character corresponding to type of plot to make (default 'all')}
}
\description{
ggplot method for starmie object
}

