% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpak.R
\name{clumpak}
\alias{clumpak}
\title{Run the CLUMPP algorithms. 
If you want to run in parallel, you'll need to spin up multiple sessions or cores using plan(multisession) or plan(multicore) (the latter won't work on Windows).}
\usage{
clumpak(Q_list, method = "none", parallel = FALSE, iter = 100)
}
\arguments{
\item{Q_list}{A list of of Q matrices.}

\item{method}{The method the algorithm uses to infer the correct permutations. One of 'greedy' or 'greedyLargeK' or 'stephens' or 'none'}

\item{parallel}{Use future.apply to parallelise clumpp?}

\item{iter}{The number of iterations to use if running either 'greedy' or 'greedyLargeK'.}
}
\description{
Run the CLUMPP algorithms. 
If you want to run in parallel, you'll need to spin up multiple sessions or cores using plan(multisession) or plan(multicore) (the latter won't work on Windows).
}
\examples{
# Read in Structure files
multiple_runs_k10 <- exampleStructure("mcmc_diagnostics")
Q_list <- lapply(multiple_runs_k10, getQ)
clumpak_results <- clumpak(Q_list)
plan(multisession,workers=4)
clumpak_results_parallel <- clumpak(Q_list, parallel=TRUE)
plan(sequential) # it's usually a good idea to go back to sequential to avoid zombie processes.
}
