% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpp_future.R
\name{clumpp_future}
\alias{clumpp_future}
\title{Run the CLUMPP algorithms faster! You'll need to spin up multiple sessions or cores using plan(multisession) or plan(multicore) (the latter won't work on Windows).}
\usage{
clumpp_future(Q_list, method = "greedy", iter = 100)
}
\arguments{
\item{Q_list}{A list of of Q matrices.}

\item{method}{The algorithm to use to infer the correct permutations. One of 'greedy' or 'greedyLargeK' or 'stephens'}

\item{iter}{The number of iterations to use if running either 'greedy' or 'greedyLargeK'}
}
\description{
Run the CLUMPP algorithms faster! You'll need to spin up multiple sessions or cores using plan(multisession) or plan(multicore) (the latter won't work on Windows).
}
\examples{
# use multiple K=3 runs
cl_data <- exampleStructure("clumpp")
print(cl_data)
Q_list <- lapply(cl_data, getQ)
plan(multisession)
clumppy <- clumpp_future(Q_list)
# Alternatively, supply parallel=TRUE to clumpak(). See ?clumpak.
}
