% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{exampleStructure}
\alias{exampleStructure}
\title{Example structure objects}
\usage{
exampleStructure(example_type)
}
\arguments{
\item{example_type}{a character string either "multiple_runs", "clumpp" or
"mcmc_diagnostics" or "barplot"}
}
\description{
load structure objects for different starmie functions
}

