% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineStruct.R
\name{fineStruct}
\alias{fineStruct}
\title{Constructor for fineStruct object}
\usage{
fineStruct()
}
\value{
an fineStruct object which is a list with 3 elements:
 nsamples: number of samples used

 cfactor: estimate of the c parameter in the fineStructure model

 chunkcounts_matrix: a data.matrix of chunkcounts

 dendro: a dendrogram object estimated by fineStructure

 mcmc_df: a data.frame containing the mcmc output of fineStructure

 meancoincedence_matrix: a data.matrix of pairwise coincidence, showing the MAP state

 mappopchunk_matrix: a data.matrix of population-by-population chunkcounts
}
\description{
Constructor for fineStruct object
}

