% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmie-methods.R
\name{fortify.starmie}
\alias{fortify.starmie}
\title{Return a tidy data frame from a starmie object for custom plotting}
\usage{
\method{fortify}{starmie}(starmie_obj, method)
}
\description{
Return a tidy data frame from a starmie object for custom plotting
}

