% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starmie.R
\name{get_ploidy}
\alias{get_admixture}
\alias{get_markers}
\alias{get_nmarkers}
\alias{get_nsamples}
\alias{get_ploidy}
\alias{get_samples}
\alias{get_structure}
\title{Accessor methods for starmie objects}
\usage{
get_ploidy(starmie_obj)

get_nsamples(starmie_obj)

get_nmarkers(starmie_obj)

get_samples(starmie_obj)

get_markers(starmie_obj)

get_structure(starmie_obj)

get_admixture(starmie_obj)
}
\description{
Return elements from a \code{\link{starmie_obj}}
}
\section{Functions}{
\itemize{
\item \code{get_nsamples}: Return number of samples

\item \code{get_nmarkers}: Return number of markers

\item \code{get_samples}: Return sample data frame

\item \code{get_markers}: Return marker data frame

\item \code{get_structure}: Return structure run list

\item \code{get_admixture}: Return admixture run list
}}

