% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getClusterAlleleFreqMat}
\alias{getClusterAlleleFreqMat}
\title{Retrieve estimated within-cluster allele frequencies}
\usage{
getClusterAlleleFreqMat(x)
}
\arguments{
\item{x}{a \code{\link{struct}} or \code{\link{admix}} object.}
}
\description{
Retrieve estimated within-cluster allele frequencies
}

