% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getCompleteAlleleFreqMat}
\alias{getCompleteAlleleFreqMat}
\title{Retrieve estimated population allele frequencies}
\usage{
getCompleteAlleleFreqMat(x)
}
\arguments{
\item{x}{a \code{\link{struct}} or \code{\link{admix}} object.}
}
\description{
Retrieve estimated population allele frequencies
}

