% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{getD}
\alias{getD}
\alias{getFitStats}
\alias{getMCMC}
\alias{getPosterior}
\title{Accessor methods for struct objects}
\usage{
getD(structure_obj)

getPosterior(structure_obj)

getFitStats(structure_obj)

getMCMC(structure_obj)
}
\arguments{
\item{structure_obj}{a \code{\link{struct}} object}
}
\description{
getD Return the number of free parameters in STRUCTURE model
}
\section{Functions}{
\itemize{
\item \code{getPosterior}: Return the estimated log posterior probability (L_k) from a \code{\link{struct}} object

\item \code{getFitStats}: Return the estimated mean and variance of estimated log-likelihood from a \code{\link{struct}} object

\item \code{getMCMC}: Return non-burn in MCMC iterations.
}}

