% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{getMCMC}
\alias{getMCMC}
\title{Acessor methods for structure object}
\usage{
getMCMC(structure_obj)
}
\description{
Acessor methods for structure object
}
\section{Functions}{
\itemize{
\item \code{getMCMC}: Return non-burn in MCMC iterations.
}}

