% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpp.R
\name{getStephens}
\alias{getStephens}
\title{Use the Stephen's method to permute sample labels}
\usage{
getStephens(Q_list)
}
\arguments{
\item{Q_list}{A list of of Q matrices.}
}
\description{
Use the Stephen's method to permute sample labels
}

