% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFineStructure.R
\name{loadFineStructure}
\alias{loadFineStructure}
\title{Read fineSTRUCTURE Output}
\usage{
loadFineStructure(chunkfile, treefile, mcmcfile, meancoincidencefile = NULL,
  mappopchunkfile = NULL)
}
\arguments{
\item{chunkfile}{a string containing a chunk counts file produce by fineStructure}

\item{treefile}{a string containing a tree file produce by fineStructure}

\item{mcmcfile}{a string containing a mcmc file produce by fineStructure}

\item{meancoincidencefile}{(default=NULL) a string containing a pairwise coincidence produce by fineStructure}

\item{mappopchunkfile}{(default=NULL) a string containing a population level chunk counts file produce by fineStructure}
}
\description{
Read fineSTRUCTURE Output
}
\examples{
# read in example fineSTRUCTURE output
chunkfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.chunkcounts.out", package = "starmie")
treefile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.tree.xml", package = "starmie")
mcmcfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.mcmc.xml", package = "starmie")
meancoincidencefile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.meancoincidence.csv", package = "starmie")
mappopchunkfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.mapstate.csv", package = "starmie")
fineData <- loadFineStructure(chunkfile, treefile, mcmcfile, meancoincidencefile, mappopchunkfile)
fineData
}

