% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaData.R
\name{loadSampleData}
\alias{loadSampleData}
\title{Load sample metadata}
\usage{
loadSampleData(starmie_obj, sample_file, pop_identifier = FALSE)
}
\arguments{
\item{sample_file}{filename containing sample information}

\item{pop_identifer}{logical does file contain geographic location?}
}
\description{
Read in sample information from space delimted PLINK fam file
or plain text file.
}
\details{
Assume that we have PLINK format FAM file with optional location identifier in the final column
}

