% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseSimilarityMatrix.R
\name{pairwiseSimilarityMatrix}
\alias{pairwiseSimilarityMatrix}
\title{Compute matrix of pairwise similarity between Q matrices.}
\usage{
pairwiseSimilarityMatrix(Q_list)
}
\arguments{
\item{Q_list}{A list of of Q matrices.}
}
\description{
Compute matrix of pairwise similarity between Q matrices.
}
\details{
Implementation of the pairwise similarity metric as
defined in Jakobsson, M. and Rosenberg, N. A., 2007.
}
\examples{
# Read in Structure files
multiple_runs_k10 <- exampleStructure("mcmc_diagnostics")
Q_list <- lapply(multiple_runs_k10, getQ)
H <- pairwiseSimilarityMatrix(Q_list)
}
