% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{Generate a barplot of a Structure or Admixture run.}
\usage{
plotBar(cluster_run, populations = NULL, plot = TRUE, facet = TRUE)
}
\arguments{
\item{cluster_run}{a single cluster run object of type \code{\link{struct}} or \code{\link{admix}} or data.frame for Q in the same format as that generated by loadAdmixture.}

\item{populations}{a data.frame that contains the sample number as the first column and the population as the second.}

\item{plot}{if FALSE returns a data.frame for customised plots}

\item{facet}{whether or not to split the barplot by cluster. This is recommended.}
}
\description{
Generate a barplot of a Structure or Admixture run.
}
\examples{
# Read in Structure files
structure_files <- system.file("extdata/microsat_testfiles", package="starmie")
structure_output_files <- list.files(structure_files, pattern = "*.out_f", full.names = TRUE)
cluster_run <- loadStructure(structure_output_files[[6]])
# Generate barplot
plotBar(cluster_run)

}

