% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFineStructDendro.R
\name{plotFineStructDendro}
\alias{plotFineStructDendro}
\title{Plot dendrogram from output of fineStructure}
\usage{
plotFineStructDendro(x, cex = 0.7, ...)
}
\arguments{
\item{x}{a \code{\link{fineStruct}} or ape::phylo object}

\item{cex}{(default=0.7) number indicating the amount by which plotting text and symbols should be scaled relative to the default.}

\item{...}{Additional arguments to be passed to ape::plot.phylo}
}
\description{
Plot dendrogram from output of fineStructure
}
\examples{
# fineStruct example
chunkfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.chunkcounts.out", package = "starmie")
treefile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.tree.xml", package = "starmie")
mcmcfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.mcmc.xml", package = "starmie")
fineData <- loadFineStructure(chunkfile, treefile, mcmcfile)
plotFineStructDendro(fineData)
}

