% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMC.R
\name{plotMCMC}
\alias{plotMCMC}
\title{Plot STRUCTURE MCMC chains}
\usage{
plotMCMC(struct_list, make_plot = TRUE, use_logL = TRUE, facet = TRUE)
}
\arguments{
\item{struct_list}{a list \code{\link{struct}} objects}

\item{make_plot}{logical print resulting plot default TRUE}

\item{use_logL}{logical plot log-likelihood (TRUE) or admixture coeffecient}

\item{facet}{logical facet by K default TRUE}
}
\value{
If make_plot is TRUE a ggplot is printed to the screen and the
plot object and the data to generate it are returned. Otherwise,
a data.frame containing MCMC info it returned.
}
\description{
Plot non-burn MCMC iterations of STRUCTURE for checking convergence.
If make_plot is set to FALSE a data.frame is returned containing the log likelihood
and alpha values over different K and runs and not plot is printed to the device.
}
\examples{
#Read in Structure files
structure_files <- system.file("extdata/microsat_testfiles", package="starmie")
structure_output_files <- list.files(structure_files, pattern = "*.out_f", full.names = TRUE)
structure_log_files <- list.files(structure_files, pattern = ".*log", full.names = TRUE)
runs <- mapply(loadStructure, structure_output_files, structure_log_files, SIMPLIFY=FALSE)
}

