% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{Plot prinicipal compenents between inferred clusters}
\usage{
plotMDS(structure_obj, method = "nnd")
}
\arguments{
\item{structure_obj}{a \code{\link{struct}} object}

\item{method}{one of "nnd" or "jsd"}
}
\description{
Plot prinicipal compenents between inferred clusters
}
\details{
"nnd" uses the nucleotide distance matrix estimated by STRUCTURE
to construct the principal coordinates, sizing the points by the expected
heterozygosity within a cluster. "jsd" produces a principal coordinates
from the Jensen Shannon Divergence metric as used by the 'ldavis' package.
}

