% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiK.R
\name{plotMultiK}
\alias{plotMultiK}
\title{Generate a barplot of a Structure or Admixture run.}
\usage{
plotMultiK(cluster_runs, populations = NULL, plot = TRUE)
}
\arguments{
\item{cluster_runs}{a list multiple cluster runs of type \code{\link{struct}} or \code{\link{admix}} or matrix for Q in the same format as that generated by loadAdmixture.}

\item{populations}{a data.frame that contains the sample number as the first column and the population as the second.}

\item{plot}{if FALSE returns a data.frame for customised plots}
}
\description{
Generate a barplot of a Structure or Admixture run.
}
\examples{
# Read in Structure files
structure_files <- system.file("extdata/microsat_testfiles", package="starmie")
structure_output_files <- list.files(structure_files, pattern = "*.out_f", full.names = TRUE)
cluster_runs <- lapply(structure_output_files[1:6], loadStructure)
# Generate barplot
plotMultiK(cluster_runs)

}

