% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPWCoancestry.R
\name{plotPWCoancestry}
\alias{plotPWCoancestry}
\title{fineSTRUCTURE pairwise coancestry plot}
\usage{
plotPWCoancestry(x, cex = 0.5, colours = NULL, keep_ones = FALSE,
  max_coancestry = NULL)
}
\arguments{
\item{x}{a \code{\link{fineStruct}} object}

\item{cex}{(default=0.7) number indicating the amount by which plotting text and symbols should be scaled relative to the default.}

\item{colours}{a vector of colors to be passed to image. (e.g., heat.colors(n=200)[200:1])}

\item{keep_ones}{(default=FALSE) whether or not to annotate node with full support.}

\item{max_coancestry}{the maximum value to be reprsented in the heatmap. Large values will be reduced to max_coancestry.}
}
\description{
fineSTRUCTURE pairwise coancestry plot
}
\examples{
# fineStruct example
chunkfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.chunkcounts.out", package = "starmie")
treefile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.tree.xml", package = "starmie")
mcmcfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.mcmc.xml", package = "starmie")
meancoincidencefile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.meancoincidence.csv", package = "starmie")
mappopchunkfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.mapstate.csv", package = "starmie")
fineData <- loadFineStructure(chunkfile, treefile, mcmcfile, meancoincidencefile, mappopchunkfile)
plotPWCoancestry(fineData, max_coancestry=500)
}

