% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPWCoincidence.R
\name{plotPWCoincidence}
\alias{plotPWCoincidence}
\title{fineSTRUCTURE pairwise coincidence plot}
\usage{
plotPWCoincidence(x, cex = 0.5, keep_ones = FALSE, colours = NULL)
}
\arguments{
\item{x}{a \code{\link{fineStruct}} object}

\item{cex}{(default=0.7) number indicating the amount by which plotting text and symbols should be scaled relative to the default.}

\item{keep_ones}{(default=FALSE) whether or not to annotate node with full support.}

\item{colours}{a vector of colors to be passed to image. (e.g., heat.colors(n=200)[200:1])}
}
\description{
fineSTRUCTURE pairwise coincidence plot
}
\examples{
# fineStruct example
chunkfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.chunkcounts.out", package = "starmie")
treefile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.tree.xml", package = "starmie")
mcmcfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.mcmc.xml", package = "starmie")
meancoincidencefile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.meancoincidence.csv", package = "starmie")
mappopchunkfile <- system.file("extdata/fine_structure_files", "EastAsiaSimple.EMlinked.mapstate.csv", package = "starmie")
fineData <- loadFineStructure(chunkfile, treefile, mcmcfile, meancoincidencefile, mappopchunkfile)
plotPWCoincidence(fineData)
}

