% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRanks.R
\name{plotRanks}
\alias{plotRanks}
\title{Plot allele frequencies within inferred cluster}
\usage{
plotRanks(structure_obj, cluster_num, relevance = NULL)
}
\arguments{
\item{structure_obj}{a \code{\link{struct}} object}

\item{cluster_num}{an integer value for the cluster of interest}

\item{relevance}{relevance metric applied to ranking}
}
\description{
Plot allele frequencies within inferred cluster
}
\details{
This function plots the estimated allele frequencies within a genetic
cluster and contrasts them to the overall population allele frequencies.
}
\examples{
k6_data <- exampleStructure("barplot")
plotRanks(k6_data, cluster_num = 2)
}

