% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeBar.R
\name{plotTreeBar}
\alias{plotTreeBar}
\title{Generate a barplot of a Structure or Admixture run.}
\usage{
plotTreeBar(cluster_run, facet = TRUE, dendro = NULL, cut = NULL)
}
\arguments{
\item{cluster_run}{a single cluster run object of type \code{\link{struct}} or \code{\link{admix}} or data.frame for Q in the same format as that generated by loadAdmixture.}

\item{facet}{whether or not to split the barplot by cluster. This is recommended.}

\item{dendro}{an object of class 'hclust' (defaults to hclust with average linkage)}

\item{cut}{an integer vector output by 'cutree' (defaults to cutree k=ncols(Q))}
}
\description{
Generate a barplot of a Structure or Admixture run.
}
\examples{
# Read file using K = 6 and plot results
k6_data <- exampleStructure("barplot")
print(k6_data)
# our facetted structure plot with tree
plotTreeBar(k6_data)
# standard 'structure' bar plot with tree
plotTreeBar(k6_data, facet = FALSE)
}

