% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAdmixture.R
\name{readAdmixture}
\alias{readAdmixture}
\title{Read Admixture Output}
\usage{
readAdmixture(path, k_values, snp_map, logfile = FALSE)
}
\arguments{
\item{path}{directory containing admixture run information}

\item{k_values}{integer vector of K values tried}

\item{snp_map}{name of PLINK map file for genomic coordinates}

\item{logfile}{include logfiles for each admixture run (required for diagnositc plots)}
}
\description{
Read Admixture Output
}

