% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStructure.R
\name{readStructure}
\alias{readStructure}
\title{Read Structure Output}
\usage{
readStructure(starmie_obj, filename, logfile = NULL)
}
\arguments{
\item{logfile}{include STRUCTURE logs for each run (required if wanting to do MCMC plots)}

\item{starmie}{object}

\item{location}{of the STRUCTURE output file}
}
\description{
Read Structure Output
}

