% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateQmatrix.R
\name{simulateQ}
\alias{simulateQ}
\title{Simulate random Q-matrix and permutations with replications}
\usage{
simulateQ(n, alpha, r)
}
\arguments{
\item{n}{integer number of samples}

\item{alpha}{vector of length K (number of columns)}

\item{r}{number of replicates}
}
\description{
Simulate random Q-matrix and permutations with replications
}

