% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{structList}
\alias{structList}
\title{Constructor for a structList object}
\usage{
structList(...)
}
\arguments{
\item{...}{a list of a \code{\link{struct}} objects}
}
\description{
the structList class is a conatainer for storing a collection
of struct objects.
}

