% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averagePairWiseSimilarityH.R
\name{averagePairWiseSimilarityH}
\alias{averagePairWiseSimilarityH}
\title{Average Q matrices.}
\usage{
averagePairWiseSimilarityH(Q_list)
}
\arguments{
\item{Q_list}{A list of of Q matrices.}
}
\description{
Average Q matrices.
}
\examples{
# Read in Structure files
multiple_runs_k10 <- exampleStructure("mcmc_diagnostics")
Q_list <- lapply(multiple_runs_k10, getQ)
avgQ <- averagePairWiseSimilarityH(Q_list)
}

