% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestK.R
\name{bestK}
\alias{bestK}
\title{Determine a suitable K value from multiple Structure runs}
\usage{
bestK(runs, method = "evanno", make_plot = TRUE)
}
\arguments{
\item{runs}{a list of \code{\link{struct}} objects or \code{\link{admix}} objects.}

\item{method}{the method used to calculate the best K}

\item{make_plot}{whether of not to generate diagnostic plots}
}
\description{
Determine a suitable K value from multiple Structure runs
}
\examples{
multi_K <- exampleStructure("multiple_runs")
# Run the evanno method and display diagnostic plots.
bestK(multi_K)
bestK(multi_K, "structure")
}

