% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestK.R
\name{bestK}
\alias{bestK}
\title{Determine a suitable K value from multiple Structure runs}
\usage{
bestK(runs, method = "evanno", make_plot = TRUE)
}
\arguments{
\item{runs}{a list of \code{\link{struct}} objects or \code{\link{admix}} objects.}

\item{method}{the method used to calculate the best K}

\item{make_plot}{whether of not to generate diagnostic plots}
}
\description{
Determine a suitable K value from multiple Structure runs
}
\examples{
# Read in Structure files
structure_files <- system.file("extdata/microsat_testfiles", package="starmie")
structure_output_files <- list.files(structure_files, pattern = "*.out_f", full.names = TRUE)
structure_log_files <- list.files(structure_files, pattern = ".*log", full.names = TRUE)
runs <- mapply(loadStructure, structure_output_files, structure_log_files, SIMPLIFY=FALSE)
# Run the evanno method and display diagnostic plots.
K <- bestK(runs)

}

