% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpak.R
\name{clumpak}
\alias{clumpak}
\title{Run the CLUMPP algorithms.}
\usage{
clumpak(Q_list, method = "none")
}
\arguments{
\item{Q_list}{A list of of Q matrices.}

\item{method}{The method the algorithm uses to infer the correct permutations. One of 'greedy' or 'greedyLargeK' or 'stephens' or 'none'}
}
\description{
Run the CLUMPP algorithms.
}
\examples{
# Read in Structure files
multiple_runs_k10 <- exampleStructure("mcmc_diagnostics")
Q_list <- lapply(multiple_runs_k10, getQ)
clumpak_results <- clumpak(Q_list)
}

