% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpp.R
\name{clumpp}
\alias{clumpp}
\title{Run the CLUMPP algorithms.}
\usage{
clumpp(Q_list, method = "greedy")
}
\arguments{
\item{Q_list}{A list of of Q matrices.}

\item{method}{The algorithm to use to infer the correct permutations. One of 'greedy' or 'greedyLargeK'}
}
\description{
Run the CLUMPP algorithms.
}
\examples{
# Read in Structure files
structure_files <- system.file("extdata/microsat_testfiles", package="starmie")
structure_output_files <- list.files(structure_files, pattern = ".*K10.*out_f", full.names = TRUE)
structure_log_files <- list.files(structure_files, pattern = ".*K10.*log", full.names = TRUE)
structure_runs <- mapply(loadStructure, structure_output_files, structure_log_files, SIMPLIFY=FALSE)
Q_list <- lapply(structure_runs, getQ)
clumpp(Q_list)
}

