% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpp.R
\name{clumpp}
\alias{clumpp}
\title{Run the CLUMPP algorithms.}
\usage{
clumpp(Q_list, method = "greedy", iter = 100)
}
\arguments{
\item{Q_list}{A list of of Q matrices.}

\item{method}{The algorithm to use to infer the correct permutations. One of 'greedy' or 'greedyLargeK' or 'stephens'}

\item{iter}{The number of iterations to use if running either 'greedy' or 'greedyLargeK'}
}
\description{
Run the CLUMPP algorithms.
}
\examples{
# use multiple K=3 runs
cl_data <- exampleStructure("clumpp")
print(cl_data)
Q_list <- lapply(cl_data, getQ)
clumppy <- clumpp(Q_list)
}

