% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{getK}
\alias{getD}
\alias{getFitStats}
\alias{getK}
\alias{getMCMC}
\alias{getPosterior}
\alias{getQ}
\title{Accessor methods for struct objects}
\usage{
getK(structure_obj)

getD(structure_obj)

getPosterior(structure_obj)

getFitStats(structure_obj)

getQ(structure_obj)

getMCMC(structure_obj)
}
\arguments{
\item{structure_obj}{a \code{\link{struct}} object}
}
\description{
Return elements from \code{\link{struct}} objects.
}
\section{Functions}{
\itemize{
\item \code{getD}: Return the number of free parameters in STRUCTURE model

\item \code{getPosterior}: Return the estimated log posterior probability (L_k) from a \code{\link{struct}} object

\item \code{getFitStats}: Return the estimated mean and variance of estimated log-likelihood from a \code{\link{struct}} object

\item \code{getQ}: Return the Q matrix

\item \code{getMCMC}: Return non-burn in MCMC iterations.
}}

