% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{getK}
\alias{getK}
\alias{getPosterior}
\title{Accessor methods for structure object}
\usage{
getK(structure_obj)

getPosterior(structure_obj)
}
\arguments{
\item{structure_obj}{a code{\link{struct}} object}

\item{structure_obj}{a code{\link{struct}} object}
}
\description{
Return K from a \code{\link{struct}}

Accessor methods for structure object
}
\section{Functions}{
\itemize{
\item \code{getPosterior}: Return the log posterior probability from a \code{\link{struct}}
}}

