% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getK}
\alias{getK}
\title{Retrieve the assumed number of populations from \code{\link{struct}} or
\code{\link{admix}} objects.}
\usage{
getK(x)
}
\arguments{
\item{x}{a \code{\link{struct}} object  or \code{\link{admix}} object.}
}
\description{
Retrieve the assumed number of populations from \code{\link{struct}} or
\code{\link{admix}} objects.
}

