% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{getQ}
\alias{getQ}
\title{Accessor methods for structure object}
\usage{
getQ(structure_obj)
}
\description{
Accessor methods for structure object
}
\section{Functions}{
\itemize{
\item \code{getQ}: Return the Q matrix from a \code{\link{struct}}
}}

