% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getQ}
\alias{getQ}
\title{Retrieve Q matrix from \code{\link{struct}} or \code{\link{admix}} objects.}
\usage{
getQ(x)
}
\arguments{
\item{x}{a \code{\link{struct}} or \code{\link{admix}} object.}
}
\description{
Retrieve Q matrix from \code{\link{struct}} or \code{\link{admix}} objects.
}

