% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAdmixture.R
\name{loadAdmixture}
\alias{loadAdmixture}
\title{Read Admixture Output}
\usage{
loadAdmixture(starmie_obj, file_list, logfile_list = NULL)
}
\arguments{
\item{starmie_obj}{an object of class \code{\link{starmie}} with sample_data filled}

\item{file_list}{a vector of .Q file names for admixture}

\item{logfile_list}{a vector of logfile names for each admixture run (required for diagnositc plots)}
}
\description{
Read Admixture Output
}
\examples{
my_starmie <- starmie()
# add in sample metadata
sample_file <- system.file("extdata/hapmap3_files", "hapmap3.fam", package="starmie")
my_starmie <- loadSampleData(my_starmie, sample_file)
# read in Q-files
hapmap_files <- system.file("extdata/hapmap3_files", package="starmie")
qfiles <- list.files(hapmap_files, pattern = ".Q$", full.names = TRUE)
my_starmie <- loadAdmixture(my_starmie, qfiles)

}

