% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAdmixture.R
\name{loadAdmixture}
\alias{loadAdmixture}
\title{Read Admixture Output}
\usage{
loadAdmixture(path, k_values, logfile = FALSE)
}
\arguments{
\item{path}{directory containing admixture run information}

\item{k_values}{integer vector of K values tried}

\item{logfile}{include logfiles for each admixture run (required for diagnositc plots)}
}
\description{
Read Admixture Output
}

