% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStructure.R
\name{loadStructure}
\alias{loadStructure}
\title{Read Structure Output}
\usage{
loadStructure(filename, logfile = NULL)
}
\arguments{
\item{starmie}{object}

\item{vector}{of locations for the STRUCTURE output files}

\item{vector}{of logfile locations (required if wanting to do MCMC plots)}
}
\description{
Read Structure Output
}
\examples{
# read in Structure files
structure_files <- system.file("extdata/microsat_testfiles", package="starmie")
structure_output_files <- list.files(structure_files, pattern = ".*out_f", full.names = TRUE)
structure_log_files <- list.files(structure_files, pattern = ".*log", full.names = TRUE)
my_struct <- loadStructure(structure_output_file[[1]], structure_log_files[[1]])

}

