% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiK.R
\name{plotMultiK}
\alias{plotMultiK}
\title{Generate a barplot for multiple values of K..}
\usage{
plotMultiK(x, populations = NULL, plot = TRUE)
}
\arguments{
\item{x}{a \code{\link{structList}} or \code{\link{admixList}} object or a list of Q-matrices.}

\item{populations}{a data.frame that contains the sample number as the first column and the population as the second.}

\item{plot}{if FALSE returns a data.frame for customised plots}
}
\description{
Generate a barplot for multiple values of K..
}
\examples{
cluster_runs <- exampleStructure("multiple_runs")
# Generate barplot
plotMultiK(cluster_runs[3:5])
}

