% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStructure.R
\name{readStructure}
\alias{readStructure}
\title{Read Structure Output}
\usage{
readStructure(path, n_runs, logfile = FALSE)
}
\arguments{
\item{path}{directory containing output of STRUCTURE runs}

\item{n_runs}{number of runs for each K value}

\item{logfile}{include STRUCTURE logs for each run (required if wanting to do MCMC plots)}
}
\description{
Read Structure Output
}

