% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{struct}
\alias{struct}
\title{Constructor for struct object}
\usage{
struct()
}
\value{
a \code{\link{struct}} object
}
\description{
struct object for storing structure run information
}
\details{
The \code{\link{struct}} object is a list with 11 elements:

 K: number of clusters estimated by Structure

 run_params: the run parameters given to the Structure program

 mem_df: assigned cluster membership proportions

 allele_freqs: Estimated net nucleotide distances between cluster

 avg_dist_df: Average distance between individuals

 fit_stats_df: Model fit statistics

 fst_df: Fst values

 ancest_df: Inferred ancestry of individuals

 clust_allele_list: Cluster allele frequencies

 burn_df: Burn in MCMC iteration output

 nonburn_df: Main MCMC iteration output
}
\seealso{
\code{\link{loadStructure}} for reading in STRUCTURE out_f files.

\code{\link{structList}} for manipulating multiple struct objects
}

